to finish blog section & change swagger

+++ app/utils/function.js
@@ -16,7 +16,7 @@ async function signAccessToken(userID){
             mobile: user.mobile
         };
         const option = {
-            expiresIn: "1h"
+            expiresIn: "1d"
         };
         JWT.sign(payload, ACCESS_TOKEN_SECRET_KEY, option, (err, token) =>{
                 if(err) reject(createError.InternalServerError("خطای سروری رخ داد"));
@@ -27,7 +27,6 @@ async function signAccessToken(userID){
 async function signRefreshToken(userID){
     return new Promise(async (resolve, reject) =>{
         const user = await userModel.findById(userID)
-        // console.log(typeof "" + userID)
         const payload = {
             mobile: user.mobile
         };
@@ -39,29 +38,30 @@ async function signRefreshToken(userID){
                 await redisClient.set("" + userID, token, {
                     EX: 31536000
                 })
-            resolve(token)
+            resolve(token);
         })
     })
 }
 async function verifyRefreshToken(token){
         return new Promise((resolve, reject) => {
             JWT.verify(token, REFRESH_TOKEN_SECRET_KEY, async (err, payload) =>{
-                if(err) return next(createError.Unauthorized("وارد حساب کاربری خود شوید"))
+                if(err) return reject(createError.Unauthorized("وارد حساب کاربری خود شوید"));
                 const {mobile} = payload || {};
                 const user = await userModel.findOne({mobile}, {password: 0, otp: 0});
-                if(!user) return next(createError.Unauthorized("حساب کاربری یافت نشد"));
-                console.log(user._id)
-                const refreshToken = await redisClient.get("" + user._id)
-                console.log(refreshToken)
+                if(!user) return reject(createError.Unauthorized("حساب کاربری یافت نشد"));
+                const refreshToken = await redisClient.get(`${user._id}`);
+                if(!refreshToken) reject(createError.Unauthorized("ورود مجدد به حساب کاربری انجام نشد"));
                 if(token === refreshToken) return resolve(mobile);
-                reject(createError.Unauthorized("ورود مجدد به حساب کاربری انجام نشد"))
+                reject(createError.Unauthorized("ورود مجدد به حساب کاربری انجام نشد"));
             })
         })
 }
 
 function deleteFileInPublic(fileAddress){
-    const pathFile = path.join(__dirname, "..", "..", fileAddress);
-    if (fs.existsSync(pathFile)) fs.unlinkSync(pathFile); 
+    if(fileAddress){
+        const pathFile = path.join(__dirname, "..", "..", fileAddress);
+        if (fs.existsSync(pathFile)) fs.unlinkSync(pathFile); 
+    }
 }
 
 module.exports = {
